Inherit the previous line colors when hitting an invalid log level
Really the reason for the log level of 65 in middle of an oops backtrace is printk()'s rather odd design conflicting with the marker for the instruction pointer in a raw code dump. It doesn't help that the ringbuffer gets line-wrapped:

<0>Code: 73 18 48 c7 c7 39 ae 4c 81 31 c0 e8 93 36 2e 00 31 c0 48 83 c4 10 5b 5d 41 5c 41 5d 41 5e c3 48 85 ff 74 1c 48 8b 87 00 02 00 00 <65> ff 40 04 83 3f 02 75 0c 48 8b bf f0 01 00 00 e9 f2 24 fd ff

Of course this doesn't affect on-screen presentation when the bomb explodes at first.